Implement react-modal library to fix overlay
Before this work, when our TKOut team brought together our separate
services we noticed my photo modal when pulled up, did not darken the
entire page behind it. So with some help from my partners with the idea,
I implemented the react-modal library to solve this problem. The
library comes with some built in operations and access to
hooks(modalIsOpen and setIsOpen) to apply an overlay of the entire
screen behind the modal.